refactor: rename files and add separate folders for forms
