+++ kickFlip-project/src/pages/login/loginPage.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '../components/form/LoginForm/form';
+import LoginForm from '../../components/form/LoginForm/loginForm';
 
 export default function LoginPage() {
     return <LoginForm />;
