+++ kickFlip-project/src/test/loginForm.test.tsx
@@ -1,6 +1,6 @@
 import { render, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import LoginForm from '../components/form/LoginForm/form';
+import LoginForm from '../components/form/LoginForm/loginForm';
 
 describe('LoginForm component', () => {
     it('displays error message when entering incorrect email', () => {
